fix(genexpr): corrected misapplication of non-integer bias
Found the issue with the golang port: the bias against non-integer solutions was being applied to integer solutions :P It's a wonder it *ever* found a solution.

With this fix, the thing can find solutions to crazy large solutions — like, any number from 0 to 10 million — in under a minute. It blows through iterations.