ref(genexpr): pre-allocate decodeToken slice
The number of pre-allocated decodeTokens is chosen mostly arbitrarily: it's the ChromosomeSize / 3. The upper bound to the number of tokens is ChromosomeSize, but, in practice, this rarely occurs.

This change brings the avg # allocations per frame down to ~770-1000. The per-iteration runtime has also become less erratic, staying around 0.28-0.31ms/iteration (previously, it would jump around from 0.29 to 0.35, mostly staying in the 0.31-0.33 range)