feat(genexpr): change gene alphabet to allow easier ±1 transitions
I'd noticed when running the 987654321 sim that it would often get stuck on 987654322 or 98765430. Since I was using 5 gene bits (a value I copied from the old Python version), and all digits and operators had only one gene (for a total of 14 genes), I had a lot of extra room in the gene alphabet to play with — 18 extra genes, to be exact. So, I restructured the genes to make it easier (read: one bit flip away) to transition between numbers. For some important transitions, like 0<->9, alternate genes were added for the numbers, hoping the additional frequency would lead to increased likelihood.

The results seem to be promising: almost no 987654321 sim has exceeded 100k iterations. Most are 35k and below.

I might have more luck in the future from assigning the remaining 15 unused genes.