ref(genexpr): avoid map lookups on tokenTypeOfByte hot path
This method would be called PopulationSize×ChromosomeSize times every iteration, causing the hash function to be called 1 or 2 times that amount. When profiling, this cost was around 11% of total runtime.

Some remaining big offenders are the sorting inside selectChromosomesXYZ (which is necessary, but could be reduced to a single sort and ChromosomeSize/2 copies per generation); generating random numbers (which could possibly be addressed by using a 3rd-party CSPRNG lib, and possibly by reading randomness in much larger chunks, then siphoning off that); map lookups for genes within Decode (which could be converted from a map to a tree structure, or simply an array, since there are currently only 32 possible genes); and Decode itself... which I'm not sure how to optimize, as of yet. Perhaps more detailed profiling tools could help with that.