fix(genexpr): switch to big.Floats to address prec error false positives/negatives
This change allows for arbitrary-precision floating point numbers to be used for intermediate (and final, but this isn't important) results within evaluation. It can be configured on the command-line with -precision <uint>.

Upon closer inspection of some pretty complex solutions, I noticed they were only evaluating to the target number due to floating point maths losing precision and disregarding fractional bits. In an effort to curb this (but not dissolve completely), I aimed to use float128s instead of float64s. The impact on runtime really hasn't been as large as I thought it would be, but the accuracy has definitely improved (I check results with WolframAlpha now, 'cause most calculators I use, namely IPython and Gnome Calculator, appear to rely on good ol' doubles).