feat(mandelbrot): add super basic golang mandelbrot GUI
The rendering is super slow, and I didn't get the zoom right (it's supposed to center the point you click on), but it works. The maths is based on the pseudocode from the Wikipedia article, and I don't yet understand _why_ it works. I want to experiment with different colours, but I really, really, really want to get zooming just right before I do that, or I'm just gonna get frustrated when I find a pretty part and I can't explore it.

This is also my first time working with homogeneous coordinates and affine transforms. At first, I was just manually mathsing the screen-to-mandelbrot-space scaling, but I switched to the matrices in order to apply the zooms. It took me some time to get the screen-to-mandelbrot scaling working with it, but when I got it working, it was because I took the time to understand affine transformations and a wee bit of matrix multiplication, so it felt nice to figure it out with my own logic.