fix(mandelbrot): resolve broken zoom-to-mouse; use mouse wheel now
The zoom issue was due to the zoom matrix being in unit scale, instead of in mandelbrot scale. Now, instead of using values between 0.0 and 1.0 for the zoom matrix's x and y, we first translate those values into their corresponding mandelbrot coords. I might be able to get away with avoiding the intermediate normalization, but for now, this settles my biz.

Also, I started experimenting with the colours. At the moment, she calculates the reverse number of iterations in a logarithmic scale (to the base of max # iterations), multiplies that by the highest 24-bit value (rounded to the nearest multiple of 25), and picks apart each 8 bit section into an RGB color. It's sorta psychedelic, and I don't hate it — but I would really like to display something with more depth. Also, I would love to do some kind of anti-aliasing.