feat(mandelbrot): switch to pixelgl for rendering; use colorgrad for pretty colours
The switch to pixelgl was for performance. colorgrad is perfect for my needs: it blends colours at certain steps, and returns the corresponding colour for a number in its domain. The colour scheme began as my attempt to match the iconic blue/yellow/white scheme seen in popular mandelbrot fractal images, but I played around with my own twinges.